Moved utils module into new utils sub-package retaining backwards compatibility. Moved notebook-specific functionality into own module inside utils sub-package.
